Move files from test/api to source/api according to documentation
