+++ src/engine/source/api/test/src/apiAuxiliarFunctions.hpp
@@ -3,6 +3,7 @@
 
 #include <netinet/in.h>
 #include <string>
+#include <random>
 #include <vector>
 
 #include <builder.hpp>
@@ -26,6 +27,28 @@ void inline initLogging(void)
         logging::loggingInit(logConfig);
         initialized = true;
     }
+}
+
+std::string inline getRandomNumber(unsigned int length)
+{
+    const std::string digits = "0123456789";
+    std::random_device rd;
+    std::mt19937 generator(rd());
+    std::uniform_int_distribution<int> distribution(0, digits.size() - 1);
+
+    std::string randomNumber;
+    for (unsigned int i = 0; i < length; ++i)
+    {
+        randomNumber += digits[distribution(generator)];
+    }
+
+    return randomNumber;
+}
+
+std::string inline generateRandomStringWithPrefix(const unsigned int length, const std::string& prefix)
+{
+    std::string randomNumber = getRandomNumber(length);
+    return prefix + randomNumber;
 }
 
 // TODO: Remove this when the builder mock is implemented
