+++ src/engine/source/api/test/src/component/test/session_test.cpp
@@ -1,7 +1,7 @@
 #include <api/test/sessionManager.hpp>
 #include <gtest/gtest.h>
 
-#include "../apiAuxiliarFunctions.hpp"
+#include "../../apiAuxiliarFunctions.hpp"
 
 using namespace api::sessionManager;
