+++ src/engine/test/CMakeLists.txt
@@ -17,7 +17,6 @@ include(GoogleTest)
 
 add_subdirectory(${TEST_SOURCE_DIR}/mocks)
 add_subdirectory(${TEST_SOURCE_DIR}/base)
-add_subdirectory(${TEST_SOURCE_DIR}/api)
 add_subdirectory(${TEST_SOURCE_DIR}/defs)
 add_subdirectory(${TEST_SOURCE_DIR}/schemf)
 add_subdirectory(${TEST_SOURCE_DIR}/builder)
